+++ beam-core/src/test/java/org/esa/beam/framework/dataop/resamp/NearestNeighbourResamplingTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2010 Brockmann Consult GmbH (info@brockmann-consult.de)
+ * Copyright (C) 2014 Brockmann Consult GmbH (info@brockmann-consult.de)
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the Free
@@ -37,20 +37,20 @@ public class NearestNeighbourResamplingTest extends TestCase {
 
     public void testComputeIndexAndGetSample() throws Exception {
         final Resampling.Index index = resampling.createIndex();
-        test(index, 0.5f, 0.0f, 0, 0, 10f);
-        test(index, 0.5f, 2.0f, 0, 2, 10f);
-        test(index, 4.5f, 0.0f, 4, 0, 50f);
-        test(index, 0.5f, 3.9f, 0, 3, 20f);
-        test(index, 2.5f, 1.0f, 2, 1, 30f);
-        test(index, 4.5f, 4.0f, 4, 4, 70f);
-        test(index, 2.9f, 2.9f, 2, 2, 20f);
+        test(index, 0.5f, 0.0f, 0.0, 0.0, 10f);
+        test(index, 0.5f, 2.0f, 0.0, 2.0, 10f);
+        test(index, 4.5f, 0.0f, 4.0, 0.0, 50f);
+        test(index, 0.5f, 3.9f, 0.0, 3.0, 20f);
+        test(index, 2.5f, 1.0f, 2.0, 1.0, 30f);
+        test(index, 4.5f, 4.0f, 4.0, 4.0, 70f);
+        test(index, 2.9f, 2.9f, 2.0, 2.0, 20f);
     }
 
-    private void test(final Resampling.Index index, float x, float y, int iExp, int jExp, float sampleExp) throws Exception {
+    private void test(final Resampling.Index index, float x, float y, double iExp, double jExp, float sampleExp) throws Exception {
         resampling.computeIndex(x, y, raster.getWidth(), raster.getHeight(), index);
         assertEquals(iExp, index.i0);
         assertEquals(jExp, index.j0);
-        float sample = resampling.resample(raster, index);
+        double sample = resampling.resample(raster, index);
         assertEquals(sampleExp, sample, 1e-5f);
     }
